Decouple backend structure to separate files. Update frontend routing and README
