+++ server/exampleENV.txt
@@ -1,0 +1,11 @@
+REACT_APP_STRIPE_KEY=
+REACT_APP_STRIPE_SECRET=
+REACT_APP_COMMERCEJS_KEY=
+
+REACT_APP_FRONTEND_URL=
+REACT_APP_SERVER_URL=
+
+DATABASE_HOST=
+DATABASE_USER=
+DATABASE_SECRET=
+DATABASE=
