+++ server/server.js
@@ -1,7 +1,7 @@
 require('dotenv').config();
 const SERVER_CONFIGS = process.env.PORT || 8080
 
-const app = require('./app').func()
+const app = require('./app')
 
 app.listen(SERVER_CONFIGS, error => {
   if (error) throw error;
